requirements.txt
django-cors-headers==3.13.0
django~=3.2.7
djangorestframework==3.14.0
psycopg2==2.9.5
django-guardian==2.4.0
djangorestframework-jwt~=1.11.0